refactor(instancessapp): change location
from views/accounts to components/accounts/instances
