+++ src/components/accounts/instances/index.js
@@ -1,1 +1,2 @@
 export * from './tabs';
+export * from './InstancesApp';
