+++ src/pages/instances/[zuid]/users.js
@@ -13,7 +13,7 @@ import { Container } from '@mui/system';
 import Main from 'layouts/Main';
 import { useZestyStore } from 'store';
 import Login from 'components/console/Login';
-import { InstancesApp } from 'views/InstancesApp/InstancesApp';
+import { InstancesApp } from 'components/accounts/instances/InstancesApp';
 import { useRouter } from 'next/router';
 import BasicTable from 'components/accounts/users/BasicTable';
 import { baseroles } from 'components/accounts/users/baseroles';
